Remove default epic code path
