--- src/errors/validationError.ts
@@ -1,3 +1,0 @@
-class ValidationError extends Error {}
-
-export { ValidationError };
