+++ src/server.test.ts
@@ -18,27 +18,8 @@ jest.mock('./api/contributionsApi', () => {
     };
 });
 
-let oldEnableDynamicTestsConfig: string | undefined;
-
-beforeEach(() => {
-    oldEnableDynamicTestsConfig = process.env.ENABLE_DYNAMIC_TESTS;
-});
-
-afterEach(() => {
-    process.env.ENABLE_DYNAMIC_TESTS = oldEnableDynamicTestsConfig;
-});
-
-const enableDynamicTests = (): void => {
-    process.env.ENABLE_DYNAMIC_TESTS = 'true';
-};
-const disableDynamicTests = (): void => {
-    process.env.ENABLE_DYNAMIC_TESTS = 'false';
-};
-
 describe('POST /epic', () => {
-    it('should return an epic when targeting is positive', async () => {
-        enableDynamicTests();
-
+    it('should return an epic', async () => {
         const { pageTracking, targeting } = testData;
 
         const res = await request(app)
@@ -59,31 +40,6 @@ describe('POST /epic', () => {
             campaignCode:
                 'gdnwb_copts_memco_2020-02-11_enviro_fossil_fuel_r2_Epic__no_article_count_Control',
             campaignId: '2020-02-11_enviro_fossil_fuel_r2_Epic__no_article_count',
-        });
-    });
-
-    it('should return the default epic when targeting is positive', async () => {
-        disableDynamicTests();
-
-        const { pageTracking, targeting } = testData;
-
-        const res = await request(app)
-            .post('/epic')
-            .send({
-                tracking: pageTracking,
-                targeting,
-            });
-
-        expect(res.status).toEqual(200);
-        expect(res.body).toHaveProperty('data');
-        expect(res.body.data).toHaveProperty('html');
-        expect(res.body.data).toHaveProperty('css');
-        expect(res.body.data).toHaveProperty('meta');
-        expect(res.body.data.meta).toEqual({
-            abTestName: 'FrontendDotcomRenderingEpic',
-            abTestVariant: 'dcr',
-            campaignCode: 'gdnwb_copts_memco_frontend_dotcom_rendering_epic_dcr',
-            campaignId: 'frontend_dotcom_rendering_epic',
         });
     });
